refactor: flatten PDP to components/pdp, delete dead product files
Move components/product/pdp/ up to components/pdp/ for a flatter
structure. Delete 12 dead legacy files from components/product/ that
were superseded by the PDP rewrite. Move recommendations.tsx and
schema.tsx into pdp/ since they're PDP-only.

components/product/ now contains only price.tsx (shared).
components/pdp/ is the single home for all PDP components.
