+++ apps/template/components/pdp/about-item.tsx
