+++ apps/template/components/pdp/auto-play-video.tsx
