+++ apps/template/components/pdp/image-gallery.tsx
