+++ apps/template/components/pdp/image-stack.tsx
