+++ apps/template/components/pdp/lightbox.tsx
