+++ apps/template/components/pdp/product-info.tsx
