+++ apps/template/components/pdp/product-media.tsx
