+++ apps/template/components/pdp/recommendations.tsx
