+++ apps/template/components/pdp/schema.tsx
