+++ apps/template/components/pdp/shop-logo.tsx
