+++ apps/template/components/pdp/specs-table.tsx
