+++ apps/template/components/pdp/variants.ts
